// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type CategoryReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "category";
};

export type AuthorReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "author";
};

export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export type BLOG_POST_QUERY_RESULT = {
  _id: string;
  title: string;
  slug: string;
  excerpt: string;
  publishedAt: string;
  readingTime: number;
  category: string;
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "blockquote" | "h2" | "h3" | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: SanityImageAssetReference;
        media?: unknown;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt: string;
        caption?: string;
        _type: "image";
        _key: string;
      }
  >;
  seo: {
    title?: string;
    description?: string;
    image?: {
      asset?: SanityImageAssetReference;
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    };
  } | null;
} | null;

// Source: ../landing-page/sanity/lib/queries.ts
// Variable: RELATED_BLOG_POSTS_QUERY
// Query: *[_type == "post" && defined(slug.current) && slug.current != $slug]    | order(publishedAt desc)[0...3] {      _id,      title,      "slug": slug.current,      excerpt,      publishedAt,      readingTime,      "category": category->title    }
export type RELATED_BLOG_POSTS_QUERY_RESULT = Array<{
  _id: string;
  title: string;
  slug: string;
  excerpt: string;
  publishedAt: string;
  readingTime: number;
  category: string;
}>;

// Source: ../landing-page/sanity/lib/queries.ts
// Variable: BLOG_POST_SLUGS_QUERY
// Query: *[_type == "post" && defined(slug.current)] {"slug": slug.current}
export type BLOG_POST_SLUGS_QUERY_RESULT = Array<{
  slug: string;
}>;